+++ Sources/WikilayerClient/WikiSummary.swift
@@ -5,20 +5,43 @@ public struct WikiSummary: Codable, Sendable, Identifiable, Equatable {
     public let id: Int64
     public let title: String
     public let urlPath: String
+    public let iconURL: URL?
+    public let pagesTree: Bool
     public let updatedAt: Date
 
     enum CodingKeys: String, CodingKey {
         case id
         case title
         case urlPath = "url_path"
+        case iconURL = "icon_url"
+        case pagesTree = "pages_tree"
         case updatedAt = "updated_at"
     }
 
-    public init(id: Int64, title: String, urlPath: String, updatedAt: Date) {
+    public init(
+        id: Int64,
+        title: String,
+        urlPath: String,
+        updatedAt: Date,
+        iconURL: URL? = nil,
+        pagesTree: Bool = false
+    ) {
         self.id = id
         self.title = title
         self.urlPath = urlPath
+        self.iconURL = iconURL
+        self.pagesTree = pagesTree
         self.updatedAt = updatedAt
+    }
+
+    public init(from decoder: any Decoder) throws {
+        let box = try decoder.container(keyedBy: CodingKeys.self)
+        id = try box.decode(Int64.self, forKey: .id)
+        title = try box.decodeIfPresent(String.self, forKey: .title) ?? ""
+        urlPath = try box.decodeIfPresent(String.self, forKey: .urlPath) ?? ""
+        iconURL = try box.decodeIfPresent(URL.self, forKey: .iconURL)
+        pagesTree = try box.decodeIfPresent(Bool.self, forKey: .pagesTree) ?? false
+        updatedAt = try box.decode(Date.self, forKey: .updatedAt)
     }
 }
 
@@ -29,6 +52,8 @@ public struct ResolvedAddress: Codable, Sendable, Equatable {
     public let language: String
     public let wikiTitle: String
     public let wikiURLPath: String
+    public let wikiIconURL: URL?
+    public let wikiPagesTree: Bool
 
     enum CodingKeys: String, CodingKey {
         case wikiID = "wiki_id"
@@ -36,6 +61,8 @@ public struct ResolvedAddress: Codable, Sendable, Equatable {
         case language
         case wikiTitle = "wiki_title"
         case wikiURLPath = "wiki_url_path"
+        case wikiIconURL = "wiki_icon_url"
+        case wikiPagesTree = "wiki_pages_tree"
     }
 
     public init(
@@ -43,13 +70,17 @@ public struct ResolvedAddress: Codable, Sendable, Equatable {
         nodeID: Int64,
         language: String = "",
         wikiTitle: String = "",
-        wikiURLPath: String = ""
+        wikiURLPath: String = "",
+        wikiIconURL: URL? = nil,
+        wikiPagesTree: Bool = false
     ) {
         self.wikiID = wikiID
         self.nodeID = nodeID
         self.language = language
         self.wikiTitle = wikiTitle
         self.wikiURLPath = wikiURLPath
+        self.wikiIconURL = wikiIconURL
+        self.wikiPagesTree = wikiPagesTree
     }
 
     public init(from decoder: any Decoder) throws {
@@ -59,6 +90,8 @@ public struct ResolvedAddress: Codable, Sendable, Equatable {
         language = try box.decodeIfPresent(String.self, forKey: .language) ?? ""
         wikiTitle = try box.decodeIfPresent(String.self, forKey: .wikiTitle) ?? ""
         wikiURLPath = try box.decodeIfPresent(String.self, forKey: .wikiURLPath) ?? ""
+        wikiIconURL = try box.decodeIfPresent(URL.self, forKey: .wikiIconURL)
+        wikiPagesTree = try box.decodeIfPresent(Bool.self, forKey: .wikiPagesTree) ?? false
     }
 }